--- dogshow/artifacts/cc-dataset-a/{{cookiecutter.artifact}}/conf/default-remotes.yaml
@@ -1,2 +1,0 @@
-main: remote1
-public: remote2
