--- dogshow/artifacts/cc-dataset-b/{{cookiecutter.artifact}}/conf/created-envs.yaml
@@ -1,2 +1,0 @@
-complete:
-  branch: main
