--- dogshow/artifacts/cc-dataset-b/{{cookiecutter.artifact}}/metadata/imported-namespaces.yaml
@@ -1,2 +1,0 @@
-dogbase:
-  uri: {{ dataset_a_repo }}
