+++ dogshow/artifacts/cc-dogracebase/{{cookiecutter.artifact}}/zimmer.yaml
@@ -1,0 +1,9 @@
+name: dogracebase
+version: v0.0
+registry: {{ test_registry }}
+imported_artifacts:
+  - dogshowbase  # can be either list or dict
+envs:
+  complete:
+    params:
+      data_root: {{ csv_path }}
