+++ dogshow/artifacts/cc-dogsuccess/{{cookiecutter.artifact}}/src/__main__.py
@@ -1,0 +1,4 @@
+from sscutils import run_step
+
+if __name__ == "__main__":
+    run_step()
