+++ dogshow/artifacts/cc-dogsuccess/{{cookiecutter.artifact}}/src/counts/__init__.py
@@ -1,0 +1,1 @@
+# this to test for a module that is a directory, not just a file
