--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/conf/default-remotes.yaml
@@ -1,1 +1,0 @@
-main: remote1
