--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/conf/project-env.yaml
@@ -1,5 +1,0 @@
-dogfirst: 
-  env: highest_comps
-  tag: public
-doglast: 
-  env: complete
