--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/metadata/imported-namespaces.yaml
@@ -1,4 +1,0 @@
-dogfirst: 
-  uri: {{ dataset_a_repo }}
-doglast: 
-  uri: {{ dataset_b_repo }}
