--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/params.yaml
@@ -1,4 +1,0 @@
-success:
-  top_status_multiplier: 10
-top_report:
-  show_top: 3
