--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/src/__init__.py
@@ -1,2 +1,0 @@
-from .pipereg import pipereg
-from . import success, get_size_counts, sex_matches, top_report
