--- dogshow/artifacts/cc-project-a/{{cookiecutter.artifact}}/src/get_size_counts/__init__.py
@@ -1,2 +1,0 @@
-# this to test for a module that is a directory, not just a file
-from .step import count_step  # noqa: F401
