--- dogshow/artifacts/cc-project-b/{{cookiecutter.artifact}}/conf/project-env.yaml
@@ -1,4 +1,0 @@
-dogbase: 
-  env: complete
-dogsuccess: 
-  env: main
