--- dogshow/artifacts/cc-project-b/{{cookiecutter.artifact}}/metadata/imported-namespaces.yaml
@@ -1,4 +1,0 @@
-dogbase: 
-  uri: {{ dataset_b_repo }}
-dogsuccess: 
-  uri: "{{ project_a_repo }}|>success"
