--- dogshow/artifacts/cc-project-b/{{cookiecutter.artifact}}/src/__init__.py
@@ -1,2 +1,0 @@
-from .pipereg import pipereg
-from . import joint_success
