--- dogshow/artifacts/cc-project-b/{{cookiecutter.artifact}}/src/joint_success.py
@@ -1,13 +1,0 @@
-# p1ns1.Success to both dog kinds (from ds2)
-# based on dotm and race
-
-from .imported_namespaces import dogsuccess, dogbase
-
-# here dogbase means ds2, where in ds2 dogbase means ds1
-# this conflict needs to be resolved
-
-
-class DogSuccess:
-
-    success = dogsuccess.Success
-    # based on dotm  and competitions in ds2
